requirements.txt
# Core FastAPI dependencies
fastapi==0.115.14
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Data processing
pandas==2.3.0
numpy==2.3.1
openpyxl==3.1.2
xlrd==2.0.1

# Data analysis and visualization
matplotlib==3.10.3
seaborn==0.13.2
plotly==5.17.0
scipy==1.16.0
scikit-learn==1.7.0

# AI/ML dependencies
openai==1.12.0
anthropic==0.7.7
groq==0.4.2
huggingface-hub==0.19.4
transformers==4.36.2
# torch is optional; install a compatible version separately if needed
# torch==2.1.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
weasyprint==66.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
